Fix hardcoded default template dir, use <sfile>
Instead of having a hardcoded value of ~/.vim/templates, use
expand("<sfile>") to determine the location of the plugin, and
pick the "templates" subdirectory located one level up in the
file system hierarchy. This makes the plugin to correctly find
the directory when installed as a bundle using Pathogen, and
when installed system-wide.