Change default prefix to something less ambiguous
The default prefix, "template." leads to a painfully ambiguous regex
that makes it pull in the wrong templates for the wrong filetypes. For
example; by default "template.m" is included in the global templates.
When we strip the prefix to check for matching template files, we get
stuck with the only search criteria being the letter "m", so we get
stuck with it pulling in files like "foo.vim" that are completely
unrelated.

The other alternative to fixing this would be setting the prefix to
something without a dot in the front so the dot gets included in the
template matching criteria. However, this means you'd have to specify
the pattern for the template right after the template's filename, which
would lead to weird and potentially misleading filenames like
"templatefoo.c", unless you were to stick solely to specifying file
extensions as patterns.

I've chosen this prefix by default because it has the following
advantages:
  - It starts with a period, so it won't clutter directory listings
	since it will be hidden by default default (which is what I think
	most people like by default)
  - The pattern you wish to match begins after a colon. Colons are rare
	in filenames, and their punctuational meaning is clear to most
	people, and it makes it a lot easier to visually separate the regex
	in the filename from the prefix.

The only downside to this is that it doesn't work in Windows systems by
default, but if you managed to get gvim installed and working properly
on a Windows system you're probably smart enough to know to look in the
documentation and change it yourself.