Refactor TLoad() and merge it with TLoadCmd()
This simplifies the code by merging the common code path shared by
the TLoad() and TLoadCmd() functions, with TLoad() now supporting
all the knobs. The [auto]commands now all just call TLoad() with
the appropriate set of arguments. Note that the "optional" template
parameter is now the last one (it seems good style to put optionals
at the end of the lost), which also makes the expansion of <args>
a bit more understandable (IMHO).

Last but not least, use -nargs=? (zero or one) instead of -nargs=*
(any number of arguments, or no arguments) so it's guaranteed that
at most one argument is ever passed to the command.