Correctly detect Git repository
The current implementation of using licensee to detect a project's
license only works in the root of a project, when the LICENSE file is a
sibling to the file being created.

More often than not, when a project is in a Git repository, there is a
singular file at the root of the project that applies to all files in
all child folders too. This commit uses the Git binary to determine the
root path of the project and passes that to licensee instead of the
directory containing the file being created.

Because this behavior adds a half-second delay to (Neo)Vim opening, it
is disabled by default and can be enabled by setting
`g:templates_detect_git` to `1`.